+++ create-myproj/template-tanstack-router/rsbuild.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "@rsbuild/core";
 import { pluginReact } from "@rsbuild/plugin-react";
-import { TanStackRouterRspack } from "@tanstack/router-plugin/rspack";
+import { tanstackRouter } from "@tanstack/router-plugin/rspack";
 
 export default defineConfig({
 	plugins: [pluginReact()],
@@ -12,7 +12,18 @@ export default defineConfig({
 	},
 	tools: {
 		rspack: {
-			plugins: [TanStackRouterRspack()],
+			plugins: [tanstackRouter({
+				target: 'react',
+				autoCodeSplitting: true,
+			})],
+			module: {
+				rules: [
+					{
+						test: /\.html$/,
+						type: 'asset/source',
+					}
+				]
+			}
 		},
 	},
 });