+++ create-myproj/template-tanstack-router/src/routes/index.tsx
@@ -1,6 +1,6 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createLazyFileRoute("/")({
+export const Route = createFileRoute("/")({
 	component: Index,
 });
 
@@ -11,5 +11,5 @@ function Index() {
 				Hello world!
 			</h1>
 		</div>
-	);
+	)
 }